Extract components and display import in modal
